backend/requirements.txt
Flask
Flask-cors
python-dotenv
gunicorn
